_legacy_v1/requirements/base.txt
# requirements/base.txt
# ====================================================================
# 本番環境最小依存関係（単一真実源）
# Phase 12.5: Environment Parity & Dependency Management System
# ====================================================================

# --- 数値計算・データ処理基盤 ----------------------------------------
numpy>=1.24.0,<2.0      # 数値計算基盤・pandas_ta,joblib,scikit-learn互換性確保
pandas>=1.5.0,<2.0      # データフレーム処理・時系列データ分析・numpy互換性確保
scikit-learn==1.7.1     # 機械学習ライブラリ・CI/CD環境統一・モデル互換性確保
joblib==1.3.2           # 並列処理・モデルシリアライゼーション

# --- HTTP通信・API接続 ------------------------------------------------
requests==2.31.0        # HTTP クライアント・外部API通信
ccxt>=2.10.0            # 暗号資産取引所API・Bitbank接続・版互換性確保

# --- 設定・環境管理 ---------------------------------------------------
python-dotenv==1.0.0    # .env ファイル読み込み・環境変数管理
PyYAML==6.0.0           # YAML設定ファイル処理・production.yml読み込み

# --- リトライ・エラーハンドリング --------------------------------------
tenacity==8.2.0         # リトライ機構・API呼び出し安定化

# --- Web API フレームワーク ------------------------------------------
fastapi>=0.100.0        # FastAPI フレームワーク・ヘルスチェックAPI・互換性確保
uvicorn>=0.20.0         # ASGI サーバー・FastAPI実行環境・互換性確保

# --- テクニカル指標計算 -----------------------------------------------
pandas-ta>=0.3.14b0     # テクニカル指標計算・TA-Lib代替・本番必須

# --- 機械学習モデル（本番環境必須）-----------------------------------
lightgbm>=4.0.0         # LightGBM・アンサンブル学習・本番推論必須
xgboost>=1.7.0          # XGBoost・アンサンブル学習・本番推論必須
optuna>=3.0.0           # ハイパーパラメータ最適化・optimizer.py必須